Revert "Revert "Close #1316""
This reverts commit d72f6a1f00b23e9f423c7b9c162af42e279b0a31.

+++ src/main/java/vazkii/quark/world/world/UndergroundBiomeGenerator.java
@@ -93,7 +93,7 @@ public class UndergroundBiomeGenerator extends MultiChunkFeatureGenerator {
 		context.wallMap.keySet().forEach(pos -> info.biome.finalWallPass(world, pos));
 		context.insideList.forEach(pos -> info.biome.finalInsidePass(world, pos));
 		
-		if(info.biome.hasDungeon() && world instanceof WorldServer && random.nextFloat() < info.biome.dungeonChance) {
+		if(info.biome.hasDungeon() && world instanceof WorldServer && random.nextDouble() < info.biome.dungeonChance) {
 			List<BlockPos> candidates = new ArrayList<>(context.wallMap.keySet());
 			candidates.removeIf(pos -> {
 				BlockPos down = pos.down();
